requirements.txt
click==8.1.7
PyYAML==6.0.2
rich==13.9.2
Jinja2==3.1.4
paramiko==3.4.0
Flask==3.0.3
